+++ packages/terra-form-select/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 * Changed
   * Removed aria-checked to make screen readers announce Selected in `terra-form-select`.
-  * Updated focus to shift back on select upon key press or mouse click event while using Safari browser in `terra-form-select`.
+  * Updated focus to shift back on `select` upon key press or mouse click event on `terra-form-single-select` in Safari browser.
 
 ## Unreleased